fix(utils): use int32 for range

+++ src/utils_mtf.py
@@ -309,8 +309,8 @@ def missing_dims(self: ALL_SHAPES, other: ALL_SHAPES):
 
 def compare_range(params: ModelParameter, dim0: mtf.Dimension, dim1: mtf.Dimension, comparison: typing.Callable):
     with tf1.variable_scope(f"compare{dim0.name}_{dim1.name}"):
-        return cast(comparison(mtf_range(params.mesh, dim0, tf.bfloat16),
-                               mtf_range(params.mesh, dim1, tf.bfloat16)),
+        return cast(comparison(mtf_range(params.mesh, dim0, tf.int32),
+                               mtf_range(params.mesh, dim1, tf.int32)),
                     params.variable_dtype.activation_dtype)
